refactor(setup): de/activate extension with browser action (#52)
When the extension icon is clicked, the extension will be added to all
tabs in minimized mode (the state will persist across them). Clicking
again will destroy the extension from all tabs. The extension will also
be re-added on refresh or URL change and added to any new tabs created.